requirements.txt
asgiref==3.5.2
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.0.12
distlib==0.3.4
Django==3.2.13
django-cors-headers==3.13.0
django-currentuser==0.5.3
django-filter==22.1
django-phonenumber-field==6.3.0
django-sequences==2.7
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
filelock==3.7.1
idna==3.3
inflection==0.5.1
jsonschema==4.6.1
phonenumbers==8.12.50
Pillow==9.1.1
platformdirs==2.5.2
PyJWT==2.4.0
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.15.0
virtualenvwrapper-win==1.2.7
